split plots container so it can read context
