+++ src/vs/workbench/contrib/positronPlots/browser/positronPlots.tsx
@@ -13,10 +13,10 @@ import { IContextMenuService } from 'vs/platform/contextview/browser/contextView
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IWorkbenchLayoutService } from 'vs/workbench/services/layout/browser/layoutService';
 import { PositronPlotsServices } from 'vs/workbench/contrib/positronPlots/browser/positronPlotsState';
-import { PositronPlotsContextProvider, usePositronPlotsContext } from 'vs/workbench/contrib/positronPlots/browser/positronPlotsContext';
+import { PositronPlotsContextProvider } from 'vs/workbench/contrib/positronPlots/browser/positronPlotsContext';
 import { IPositronPlotsService } from 'vs/workbench/services/positronPlots/common/positronPlots';
 import { DisposableStore } from 'vs/base/common/lifecycle';
-import { PlotInstance } from 'vs/workbench/contrib/positronPlots/browser/components/plotInstance';
+import { PlotsContainer } from 'vs/workbench/contrib/positronPlots/browser/components/plotsContainer';
 
 /**
  * PositronPlotsProps interface.
@@ -44,8 +44,6 @@ export const PositronPlots = (props: PropsWithChildren<PositronPlotsProps>) => {
 	const [width, setWidth] = useState(props.reactComponentContainer.width);
 	const [height, setHeight] = useState(props.reactComponentContainer.height);
 
-	const positronPlotsContext = usePositronPlotsContext();
-
 	// Add IReactComponentContainer event handlers.
 	useEffect(() => {
 		// Create the disposable store for cleanup.
@@ -64,17 +62,9 @@ export const PositronPlots = (props: PropsWithChildren<PositronPlotsProps>) => {
 	// Render.
 	return (
 		<PositronPlotsContextProvider {...props}>
-			<div className='positron-plots'>
-				{positronPlotsContext.positronPlotInstances.length === 0 &&
-					<span>Plot container: {height} x {width}</span>}
-				{positronPlotsContext.positronPlotInstances.map((plotInstance, _index) => (
-					<PlotInstance
-						key={plotInstance.id}
-						width={width}
-						height={height}
-						plotClient={plotInstance} />
-				))}
-			</div>
+			<PlotsContainer
+				width={width}
+				height={height} />
 		</PositronPlotsContextProvider>
 	);
